move user data transformation from frontend to backend
